Work in progress.

+++ OCA/chapter-3/oca-sierra/primitive-casting/src/main/java/com/foo/SomeClass.java
@@ -3,7 +3,7 @@ package com.foo;
 public class SomeClass {
     private byte a = 42;
 
-    public double getA() {
+    public float getA() {
         return a;
     }
 }
